requirements.txt
streamlit==1.36.0
pandas==2.2.2
python-docx
PyPDF2